extract private methods from chatCurlClient
